fix: cache track metadata ourselves; save bitrate, duration from parse
This should prevent the metadata displayed in the UI from changing to
that of the next track before the current track has finished.

We were caching using $song->pluginData(), but this must get overwritten
the next time getNextTrack() is called.  $song->streamUrl, however, still
reflects the currently playing track, so this change implements a
metadata cache keyed on the streamUrl.

Also, we use the bitrate and duration determined by parseRemoteHeader().
Advertisement data from the service does not include duration, so
capturing it from parseRemoteHeader() helps to maintain the correct
metadata display.

Fixes #30