fix: ignore failed get shuffle station api call
If the user's station list is empty, the get shuffle station api call will
fail with HTTP status 400.  This will cause us to display a somewhat
cryptic error message in the UI.  With this change, we swallow the error,
and the UI will instead show an empty list.

We could get the station list first and then only request the shuffle
station if the station list is not empty, but for now we'll prefer to
issue the api calls in parallel as it seems slightly faster.

Fixes #32