fix: use custom station play action to play the correct station
When the user selects a station to play, LMS asks Pyrrha for a fresh station
list and then uses the station metadata at the same index in the new list as
that of the selected station in the old list.  But, the order of stations in
the new list may have changed, so LMS can end up playing the wrong station.

To deal with this, we define our own play action in which we embed the
static station url.  When LMS requests a fresh station list for the purposes
of starting playback, we can use the data from our custom play action to
select the correct station even if the list has changed.

Fixes #39